back-end/requirements.txt
pdfplumber>=0.10.0
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
